Adds the useof <inttypes.h> to adjust the size of pointers on a 64-bit system. Also added a check in libsimularities ncd() to verify that s2 was not referencing a null pointer.

+++ classification/libsimilarity/similarity.h
@@ -25,8 +25,8 @@ struct libsimilarity {
    void *cmp;
    unsigned size_cmp;
 
-   unsigned int *corig;
-   unsigned int *ccmp;
+   size_t *corig;
+   size_t *ccmp;
 
    float res;
 };
@@ -40,7 +40,7 @@ extern "C" {
 }
 #else
 void set_compress_type(int);
-unsigned int compress(int, void *, unsigned int);
+unsigned int compress(int, void *, size_t);
 int ncd(int, libsimilarity_t *);
 int ncs(int, libsimilarity_t *);
 int cmid(int, libsimilarity_t *);
